drivers: video_mcux_isi: add barcode_scanner workaround
ISI is having issues with parallel CSI when CSC is enabled. If input
format height is set to 720 for a 1280x720 camera (MT9M114), ISI will
output a unchanged frame buffer every 2 frames. Setting input format
height to a number greater than 720 (722 here) solves the issue.

This also enables cropping, alpha insertion and hflip for
barcode_scanner demo.

This is a temporary workaround.

Signed-off-by: Zoey Li <zoey.li@nxp.com>