implemented ordinal dependency resolution
implemented ordinal dependency resolution support- this allows us to
reference clock children by their ordinals in the first link stage, so
that the children clock references are discarded. Then in the second
link stage, we fill in references to child nodes with actual pointers,
so that the final implementation has correct pointers but can
optimize out what is not needed.

Signed-off-by: Daniel DeGrasse <daniel.degrasse@nxp.com>