drivers: display: mcux_elcdif: rework frame completion logic
Rework the frame completion logic for the eLCDIF. Per the documentation
of the eLCDIF frame completion interrupt, the interrupt will be asserted
at every vertical blanking interval. In order to be sure the eLCDIF has
streamed a full frame we need to:
- wait for the start of a vertical blanking interval
- wait for the next vertical blanking interval, as that indicates a full
  frame was rendered

In order to do this, wait for two frame completion interrupts from the
eLCDIF every time we reload the framebuffer. The first interrupt signals
that our new frame is now being displayed, the second interrupt signals
it is fully rendered and the buffer is no longer needed by the eLCDIF.

Fixes #80590

Signed-off-by: Daniel DeGrasse <daniel.degrasse@nxp.com>