spi_nxp_lpspi: Fix faulting control reg access
On some of the platforms, the module doesn't get clocked until the
SDK Init call, causing the control register (CR) write to fault.

The reason for this code in the first place was to avoid S32 chip errata
that caused fifos to need internal logic reset. Since we want to avoid
reconfiguring, and a module reset would require that, there is one other
documented workaround which is to just read the RX fifo register to
update the internal hardware pointer. For the TX fifo the fix is to wait
for transfer completion which should be done by the point of starting a
new one.

Signed-off-by: Declan Snyder <declan.snyder@nxp.com>