spi_nxp_lpspi: Reintroduce fast-path to code
No need to reconfigure SPI if already configured with same
configuration. This reduces latency of some transfers that
re-use the same spi_cfg.

Also, a slightly related change is to remove the setting of the CONTC
bit in the init. This was a mistake to have because there was no
intention to do a continuing command and could cause some problem, or at
least cause confusion to readers of the driver code.

Signed-off-by: Declan Snyder <declan.snyder@nxp.com>
Co-authored-by: Raymond Lei <raymond.lei@nxp.com>