samples: i2s_codec: Fix buffer size
Use BLOCK_SIZE instead of the test 16 kHz sine wave length.

This accounts for situations where the sample queue can't be big enough,
mainly when it doesn't fit in the device's SRAM and the count of samples
queued will be less. As the test sine wave vector contains ~42 periods,
so it can be truncated.

Signed-off-by: Vit Stanicek <vit.stanicek@nxp.com>