drivers: Counter: Fix LPTMR prescaler setting not work issue
In current code, if timer-mode-sel is 0(Timer Counter Mode),
bypass_prescaler_glitch will be 1, that makes prescaler setting
be bypassed. But this setting is very useful, especially for
long timing count.

According to RM, prescaler setting ranges from 2,4,...,65536,
so we can infer that when prescaler setting is not 1, we should
make bypass_prescaler_glitch be false, which means the clock is
from the output of the prescaler.

Signed-off-by: Felix Wang <fei.wang_3@nxp.com>