net: l2: openthread: improve border router packet forwarding logic
There are some cases when openthread opens a sockets and doesn't choose
as default interface OpenThread interface, this leading to usage of
platform UDP module which will then send back the packet to the
OpenThread interface. In this case, the packet should not be treated as
originated from backbone interface.

Signed-off-by: Cristian Bulacu <cristian.bulacu@nxp.com>