samples: subsys: gpu: vglite_examples: Enable vglite example execution on RT700 platform
Add support for running the vglite example applications on the NXP
RT700 platform. This includes the updates required to execute the tile,
tiger, clock, toolkit, vector, and cube samples using the gpu-vglite
subsystem.

The changes update RT700 GPU configuration, memory settings, and the
necessary compatibility updates to ensure correct execution of the
vglite_examples on RT700 in line with other supported platforms.

Signed-off-by: Harikrushna Parmar <harikrushna.parmar@nxp.com>