boards: nxp: change flag 'XIP_BOOT_HEADER_DCD_ENABLE' to local scope
This flag is used in:
1. boards/nxp/rt10xx
2. boards/nxp/rt11xx
3. hal_nxp/rt10xx/fsl_flexspi_nor_boot.h
4. hal_nxp/rt11xx/fsl_flexspi_nor_boot.h

for case1 and 2, we just leave it in boards CMakeLists.txt but
convert it to use zephyr_library_compile_definitions() to limit
its scope.

for case3 and 4, this flag is used in hal_nxp fsl_flexspi_nor_boot.h,
which is included by soc.c, therefore, this flag should be passed
in the SoC layer CMake using zephyr_library_compile_definitions().

Signed-off-by: Zhaoxiang Jin <Zhaoxiang.Jin_1@nxp.com>