openthread: platform: infra_if: Remove NAT64 packet checksums
OpenThread NAT64 translator creates a packet that has header checksums
computed and appended.
In case of using ethernet as backbone interface, some NIC may want to
have checksums set to 0 for a correct computation.
This commit aims to address this issue by checking if hardware has
offload capabilities and setting NAT64 packet checksums to 0 if needed

Signed-off-by: Cristian Bulacu <cristian.bulacu@nxp.com>