dts: arm: nxp: add KW45B41Z SoC description
Add a SoC-level DTSI for NXP KW45B41Z.

The KW45 SoC reuses the MCXW7xx peripheral layout but does not include
an IEEE 802.15.4 radio block. CAN and FlexIO are present, while LIN is
not available on all variants.

This DTSI is intended to be reused by multiple KW45-based boards.

Signed-off-by: Sumit Batra <sumit.batra@nxp.com>