boards: nxp: add X-MR-KW45-BLE-LR board
Add support for the NXP X-MR-KW45-BLE-LR board.

The board is based on the KW45B41Z wireless MCU and targets long-range
Bluetooth LE applications. It provides on-chip flash only, a CAN
transceiver, and exposes UART, SPI, I2C, ADC and PWM peripherals.

IEEE 802.15.4 is not present on this board and is disabled at the SoC
integration level.

Signed-off-by: Sumit Batra <sumit.batra@nxp.com>